Add ability to open block settings upwards
