+++ src/styles/settings.css
@@ -1,7 +1,17 @@
 .ce-settings {
   position: absolute;
   z-index: 2;
-  top: var(--toolbar-buttons-size);
+  --gap: 8px;
+
+  @media (--not-mobile){
+    position: absolute;
+    top: calc(var(--toolbox-buttons-size) + var(--gap));
+    left: 0;
+
+    &--opened-top {
+      top: calc(-1 * (var(--gap) + var(--popover-height)));
+    }
+  }
 
   &__button {
     @apply --toolbar-button;
